splitted user into account and user
