+++ backend/controllers/accountController.js
@@ -1,0 +1,96 @@
+const jwt = require('jsonwebtoken')
+const bcrypt = require('bcryptjs')
+const asyncHandler = require('express-async-handler')
+const User = require('../models/accountModel')
+
+// @desc Register New User 
+// @route POST / api/accounts/register 
+// @access Public 
+const registerUser = asyncHandler(async (req, res) => {
+    const {username, email, password} = req.body 
+    if(!username || !email || !password) {
+        res.status(400)
+        throw new Error('Please add all fields')
+    }
+
+    //Checks user existence 
+    const userExists = await User.findOne({email})
+
+    if(userExists){
+        res.status(400)
+        throw new Error('User already exists')
+    }
+
+    // Hash password
+    const salt = await bcrypt.genSalt(10)
+    const hashedPassword = await bcrypt.hash(password, salt)
+
+    // Create User
+    const user = await User.create({
+        username,
+        email,
+        password: hashedPassword
+    })
+
+    if (user) {
+        res.status(201).json({
+            _id: user.id,
+            username: user.username,
+            email: user.email,
+            token: generateToken(user._id)
+        })
+    } else{
+        res.status(400)
+        throw new Error('Invalid user data')
+    }
+
+});
+
+// @desc Authenticate a User 
+// @route POST / api/account/login
+// @access Public 
+const loginUser = asyncHandler(async (req, res) => {
+    const {email, password} = req.body
+
+    // check for user email
+    const user = await User.findOne({email})
+
+    if(user && (await bcrypt.compare(password, user.password))) {
+    res.json( { 
+        _id: user.id,
+        name: user.name,
+        email: user.email,
+        token: generateToken(user._id)
+        })
+    } else {
+        res.status(400)
+        throw new Error('Invalid Credentials')
+    }
+    res.json({ message: 'Login User' })
+});
+
+// @desc Get user data
+// @route GET / api/account/me
+// @access Private 
+const getMe = asyncHandler(async (req, res) => {
+    const {_id, name, email} = await User.findById(req.user.id)
+
+    res.status(200).json({
+        id: _id,
+        name, 
+        email,
+    })
+});
+
+//Generate JWT
+const generateToken = (id) => {
+    return jwt.sign({ id }, process.env.JWT_SECRET, {
+        expiresIn: '30d',
+    })
+};
+
+module.exports = {
+    registerUser,
+    loginUser,
+    getMe
+}
