+++ backend/middleware/authMiddleware.js
@@ -1,6 +1,6 @@
 const jwt = require('jsonwebtoken')
 const asyncHandler = require('express-async-handler')
-const User = require('../models/userModel')
+const User = require('../models/accountModel')
 
 const protect = asyncHandler(async (req,res,next) => {
     let token
