+++ backend/models/goalModel.js
@@ -6,7 +6,7 @@ const goalSchema = mongoose.Schema(
     user:{
         type: mongoose.Schema.Types.ObjectId,
         required: true,
-        ref: 'User'
+        ref: 'User_test'
     },
     text: {
         type: String,
