+++ backend/routes/userRoutes.js
@@ -1,17 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const { registerUser, loginUser, getMe, getUsers, setUser, updateUser, deleteUser, updateUserClassroom  } = require('../controllers/userController')
-
-const {protect} = require('../middleware/authMiddleware')
-
-// AUTHENTICATION CODE
-//adding a user 
-router.post('/', registerUser)
-//authenticate
-router.post('/login', loginUser)
-//get user info
-router.get('/me', protect, getMe)
-
+const { getUsers, setUser, updateUser, deleteUser, updateUserClassroom  } = require('../controllers/userController')
 
 // USER INFO CRUD 
 router.route('/').get(getUsers).post(setUser)
