+++ backend/server.js
@@ -12,6 +12,7 @@ app.use(express.urlencoded({ extended:false}))
 app.use('/api/goals', require('./routes/goalRoutes'))
 app.use('/api/users', require('./routes/userRoutes'))
 app.use('/api/subjects', require('./routes/subjectRoutes'))
+app.use('/api/accounts', require('./routes/accountRoutes'))
 app.use(errorHandler)
 
 app.listen(port, () => console.log(`Server started on port ${port}`))
